pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "gonka-tools"
version = "0.1.0"
description = "Comprehensive toolchain for Gonka.ai compute node management"
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
authors = [
    { name = "Gonka Tools" }
]
keywords = ["gonka", "mining", "gpu", "compute", "monitoring"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: System :: Monitoring",
    "Topic :: Utilities",
]

dependencies = [
    "httpx>=0.25.0",
    "aiohttp>=3.9.0",
    "python-telegram-bot>=21.0",
    "paramiko>=3.4.0",
    "fabric>=3.2.0",
    "typer>=0.9.0",
    "rich>=13.7.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "python-dotenv>=1.0.0",
    "pandas>=2.1.0",
    "tabulate>=0.9.0",
    "schedule>=1.2.0",
    "apscheduler>=3.10.0",
    "psutil>=5.9.0",
    "pyyaml>=6.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.0.0",
]

[project.scripts]
gonka = "gonka_tools.cli:cli"

[project.urls]
Homepage = "https://github.com/yourusername/gonka-tools"
Repository = "https://github.com/yourusername/gonka-tools"

[tool.hatch.build.targets.wheel]
packages = ["gonka_tools"]

[tool.black]
line-length = 100
target-version = ["py310", "py311", "py312"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "W"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_ignores = true

